EAS-2012 Clean Up DB on Functional Test Run (#102)
* first pass at purge of test data

* fix unit test

* fix query

* fix table queries

* fix query

* fix query

* fix query

* fix query

* fix query

* fix query

* fix query

* fix query

* fix query

* fix service purge

* revert change

* fix test

* test fix to service deletion

* try to avoid connection warning

* turn session sync off

* tidy up purge methods

* fix unit test

* update method name

* version bump on pytest and dependencies

* revert version changes

* add tests for purge methods

* add tests for purge methods

* fix tests for templates and folder purging

+++ app/dao/services_dao.py
@@ -381,43 +381,47 @@ def dao_remove_user_from_service(service, user):
         db.session.commit()
 
 
+@autocommit
 def delete_service_and_all_associated_db_objects(service):
-    def _delete_commit(query):
+    def _delete(query):
         query.delete(synchronize_session=False)
-        db.session.commit()
 
-    subq = db.session.query(Template.id).filter_by(service=service).subquery()
-    _delete_commit(TemplateRedacted.query.filter(TemplateRedacted.template_id.in_(subq)))
+    template_ids = db.session.query(Template.id).filter_by(service=service)
+    _delete(TemplateRedacted.query.filter(TemplateRedacted.template_id.in_(template_ids)))
 
-    _delete_commit(ServiceSmsSender.query.filter_by(service=service))
-    _delete_commit(ServiceEmailReplyTo.query.filter_by(service=service))
-    _delete_commit(ServiceLetterContact.query.filter_by(service=service))
-    _delete_commit(ServiceContactList.query.filter_by(service=service))
-    _delete_commit(InvitedUser.query.filter_by(service=service))
-    _delete_commit(Permission.query.filter_by(service=service))
-    _delete_commit(NotificationHistory.query.filter_by(service=service))
-    _delete_commit(Notification.query.filter_by(service=service))
-    _delete_commit(Job.query.filter_by(service=service))
-    _delete_commit(Template.query.filter_by(service=service))
-    _delete_commit(TemplateHistory.query.filter_by(service_id=service.id))
-    _delete_commit(ServicePermission.query.filter_by(service_id=service.id))
-    _delete_commit(ApiKey.query.filter_by(service=service))
-    _delete_commit(ApiKey.get_history_model().query.filter_by(service_id=service.id))
-    _delete_commit(AnnualBilling.query.filter_by(service_id=service.id))
+    _delete(ServiceSmsSender.query.filter_by(service=service))
+    _delete(ServiceEmailReplyTo.query.filter_by(service=service))
+    _delete(ServiceLetterContact.query.filter_by(service=service))
+    _delete(ServiceContactList.query.filter_by(service=service))
+    _delete(InvitedUser.query.filter_by(service=service))
+    _delete(Permission.query.filter_by(service=service))
+    _delete(NotificationHistory.query.filter_by(service=service))
+    _delete(Notification.query.filter_by(service=service))
+    _delete(Job.query.filter_by(service=service))
+    _delete(Template.query.filter_by(service=service))
+    _delete(TemplateHistory.query.filter_by(service_id=service.id))
+    _delete(ServicePermission.query.filter_by(service_id=service.id))
+    _delete(ApiKey.query.filter_by(service=service))
+    _delete(ApiKey.get_history_model().query.filter_by(service_id=service.id))
+    _delete(AnnualBilling.query.filter_by(service_id=service.id))
 
     verify_codes = VerifyCode.query.join(User).filter(User.id.in_([x.id for x in service.users]))
     list(map(db.session.delete, verify_codes))
-    db.session.commit()
+
+    created_by_id = Service.query.filter_by(id=service.id).one().created_by_id
     users = [x for x in service.users]
     for user in users:
-        user.organisations = []
-        service.users.remove(user)
-    _delete_commit(Service.get_history_model().query.filter_by(id=service.id))
+        if user.id != created_by_id:
+            user.organisations = []
+            service.users.remove(user)
+
+    _delete(Service.get_history_model().query.filter_by(id=service.id))
+
     db.session.delete(service)
-    db.session.commit()
+
     for user in users:
-        db.session.delete(user)
-    db.session.commit()
+        if user.id != created_by_id:
+            db.session.delete(user)
 
 
 def dao_fetch_todays_stats_for_service(service_id):
